Added the ci-cd-project

+++ sourcephase.sh
@@ -1,9 +1,15 @@
 #!/bin/bash
 
 REPO_URL="https://github.com/TaliaGeo/ci-cd-sourceRepo.git"
-CLONE_DIR=" temprory-sourceRepo"
+CLONE_DIR="temprory-sourceRepo"
 
-rm -rf "$CLONE_DIR"
-git clone $REPO_URL
+if [ -d "$CLONE_DIR" ]; then
+  echo "🧹 Removing old clone directory..."
+  rm -rf "$CLONE_DIR"
+fi
+
+
+echo "📥 Cloning repository..."
+git clone "$REPO_URL" "$CLONE_DIR"
  
  python3 cicdpipeline.py
